+++ R/biproportional-upper.R
@@ -10,11 +10,11 @@
 #'   are named. If the number of seats per district should be calculated according to the number
 #'   of votes (not the general use case), a single number for the total number of seats can be
 #'   used.
-#' @param use_list_votes By default (`TRUE`) it's assumed that each voter in a district has
+#' @param weight_votes By default (`TRUE`) it's assumed that each voter in a district has
 #'   as many votes as there are seats in a district. Thus, votes are weighted according to
-#'   the number of available district seats with [weight_list_votes()]. Set to `FALSE` if
-#'   `votes_matrix` shows the number of voters (i.e. they can only cast one vote for one
-#'   party).
+#'   the number of available district seats with [weight_votes_matrix()]. Set to `FALSE` if
+#'   `votes_matrix` shows the number of _voters_ (e.g. because they can only cast one vote for
+#'   one party).
 #' @param method Apportion method that defines how seats are assigned, see [proporz()]. Default
 #'   is the Saintë-Lague/Webster method.
 #'
@@ -37,14 +37,14 @@
 #'
 #' @export
 upper_apportionment = function(votes_matrix, district_seats,
-                               use_list_votes = TRUE,
+                               weight_votes = TRUE,
                                method = "round") {
     # check parameters
     .vmn = deparse(substitute(votes_matrix))
     .dsn = deparse(substitute(district_seats))
     votes_matrix <- prep_votes_matrix(votes_matrix, .vmn)
     district_seats <- prep_district_seats(district_seats, votes_matrix, .dsn, .vmn)
-    assert(length(use_list_votes) == 1 && is.logical(use_list_votes))
+    assert(length(weight_votes) == 1 && is.logical(weight_votes))
 
     # district seats
     if(length(district_seats) == 1) {
@@ -55,8 +55,8 @@ upper_apportionment = function(votes_matrix, district_seats,
     }
 
     # party seats
-    if(use_list_votes) {
-        votes_matrix <- weight_list_votes(votes_matrix, seats_district)
+    if(weight_votes) {
+        votes_matrix <- weight_votes_matrix(votes_matrix, seats_district)
     }
     seats_party = proporz(rowSums(votes_matrix), sum(seats_district), method)
 
@@ -73,21 +73,26 @@ upper_apportionment = function(votes_matrix, district_seats,
 #'
 #' Weight list votes by dividing the votes matrix entries by the number
 #' of seats per district. This method is used in [upper_apportionment()] if
-#' `use_list_votes` is `TRUE` (default).
+#' `weight_votes` is `TRUE` (default).
 #'
 #' @param votes_matrix votes matrix
 #' @param district_seats seats per district, vector with same length
 #'   as `ncol(votes_matrix)` and names as `colnames(votes_matrix)`
 #'
-#' @note The weighted votes are not rounded. Matrix and vector names are ignored.
+#' @note `weight_list_votes()` has been renamed to [weight_votes_matrix()]
+#' in `v1.5.2` and is deprecated.
 #'
-#' @returns the weighted `votes_matrix`
+#' @returns the weighted `votes_matrix` which contains the number of voters (not rounded)
 #'
 #' @examples
-#' weight_list_votes(uri2020$votes_matrix, uri2020$seats_vector)
+#' weight_votes_matrix(uri2020$votes_matrix, uri2020$seats_vector)
 #'
+#' @name weight_votes_matrix
+NULL
+
+#' @rdname weight_votes_matrix
 #' @export
-weight_list_votes = function(votes_matrix, district_seats) {
+weight_votes_matrix = function(votes_matrix, district_seats) {
     assert(all(district_seats >= 0))
     if(ncol(votes_matrix) != length(district_seats)) {
         stop("`length(district_seats)` must be the same as `ncol(votes_matrix)`", call. = FALSE)
@@ -104,4 +109,13 @@ weight_list_votes = function(votes_matrix, district_seats) {
     votes_matrix <- div0(votes_matrix, M_seats_district)
 
     return(votes_matrix)
+}
+
+#' @rdname weight_votes_matrix
+#' @export
+weight_list_votes = function(votes_matrix, district_seats) {
+    .Deprecated(
+        new = "weight_votes_matrix",
+        msg = "weight_list_votes has been renamed to weight_votes_matrix")
+    weight_votes_matrix(votes_matrix, district_seats)
 }