+++ R/quorum.R
@@ -10,7 +10,7 @@
 #'   [reached_quorum_any_district()].
 #' @param total Vote threshold a party must reach for all votes cast. Used as
 #'   share of total votes if less than 1. Otherwise as number of votes. Note that
-#'   votes are not weighted with [weight_list_votes()] across districts. Must be
+#'   votes are not weighted with [weight_votes_matrix()] across districts. Must be
 #'   greater than 0. Uses [reached_quorum_total()].
 #'
 #' @details There's a difference in how the functions work. With `quorum_any`,
@@ -96,7 +96,7 @@ create_quorum_function_list = function(type, any_district, total) {
 #' @note
 #' Votes are not weighted across districts. This is relevant if the quorum threshold is the
 #' minimal number of *voters* (either as percentage or absolute value). In this case, use
-#' [weight_list_votes()] before calculating the quorum.
+#' [weight_votes_matrix()] before calculating the quorum.
 #'
 #' @seealso [reached_quorum_any_district()]
 #'